+++ hcloud/resource_hcloud_server_network.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hetznercloud/hcloud-go/hcloud"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/merge"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/merge"
 )
 
 func resourceServerNetwork() *schema.Resource {