+++ internal/certificate/testing.go
@@ -3,14 +3,15 @@ package certificate
 import (
 	"context"
 	"fmt"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/testsupport"
 	"strconv"
 	"testing"
+
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/testsupport"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hetznercloud/hcloud-go/hcloud"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/testtemplate"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/testtemplate"
 )
 
 func init() {