+++ internal/datacenter/data_source_test.go
@@ -1,12 +1,13 @@
 package datacenter_test
 
 import (
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/datacenter"
 	"testing"
 
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/datacenter"
+
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/testsupport"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/testtemplate"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/testsupport"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/testtemplate"
 )
 
 func TestAccHcloudDataSourceDatacenterTest(t *testing.T) {