+++ internal/floatingip/resource_assignment_test.go
@@ -2,14 +2,15 @@ package floatingip_test
 
 import (
 	"fmt"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/floatingip"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/server"
 	"testing"
+
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/floatingip"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/server"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hetznercloud/hcloud-go/hcloud"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/testsupport"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/testtemplate"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/testsupport"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/testtemplate"
 )
 
 func TestFloatingIPAssignmentResource_Basic(t *testing.T) {