+++ internal/testsupport/dir_test.go
@@ -4,8 +4,8 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/testsupport"
 	"github.com/stretchr/testify/assert"
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/testsupport"
 )
 
 func TestProjectRoot(t *testing.T) {