+++ internal/testtemplate/dirs.go
@@ -4,7 +4,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/terraform-providers/terraform-provider-hcloud/internal/testsupport"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/testsupport"
 )
 
 // ResourceTemplateDir returns the path to the directory where the templates